Added fix for email send on throughput

+++ src/main/java/com/impetus/kundera/ycsb/runner/CassandraRunner.java
@@ -104,10 +104,10 @@ public class CassandraRunner extends YCSBRunner
     protected void sendMail()
     {
         Map<String, Double> delta = new HashMap<String, Double>();
-        double kunderaPelopsToPelopsDelta = ((timeTakenByClient.get(clients[0]) - timeTakenByClient.get(clients[1]))
-                / timeTakenByClient.get(clients[1]) * 100);
-        double kunderaThriftToThriftDelta = ((timeTakenByClient.get(clients[2]) - timeTakenByClient.get(clients[3]))
-                / timeTakenByClient.get(clients[3]) * 100);
+        double kunderaPelopsToPelopsDelta = ((timeTakenByClient.get(clients[0]).doubleValue() - timeTakenByClient.get(clients[1]).doubleValue())
+                / timeTakenByClient.get(clients[1]).doubleValue() * 100);
+        double kunderaThriftToThriftDelta = ((timeTakenByClient.get(clients[2]).doubleValue() - timeTakenByClient.get(clients[3]).doubleValue())
+                / timeTakenByClient.get(clients[3]).doubleValue() * 100);
 //        double kunderaPelopsToHectorDelta = ((timeTakenByClient.get(clients[1]) - timeTakenByClient.get(clients[4]))
 //                / timeTakenByClient.get(clients[1]) * 100);
         delta.put("kunderaPelopsToPelopsDelta", kunderaPelopsToPelopsDelta);
