+++ src/main/java/com/impetus/kundera/ycsb/runner/Neo4jRunner.java
@@ -59,8 +59,8 @@ public class Neo4jRunner extends YCSBRunner
     {
         Map<String, Double> delta = new HashMap<String, Double>();
 
-        double kunderaNativeToNativeDelta = ((timeTakenByClient.get(clients[0]) - timeTakenByClient.get(clients[1]))
-                / timeTakenByClient.get(clients[1]) * 100);
+        double kunderaNativeToNativeDelta = ((timeTakenByClient.get(clients[0]).doubleValue() - timeTakenByClient.get(clients[1]).doubleValue())
+                / timeTakenByClient.get(clients[1]).doubleValue() * 100);
         delta.put("KunderaNeo4JToNativeDelta", kunderaNativeToNativeDelta);
 
         if (kunderaNativeToNativeDelta > 8.00)
